// src/app/admin/employees/type.ts
export type SearchParams = Partial<{
  // firstName: any;
  // lastName: string;
  // communityId: string;
  // departmentIds?: string;
  keywords: string;
  roleId: string;
  license: string;
  status?: string;
  // startDate?: string[] | null;
  // endDate?: [string | null, string | null] | null;
  // exceptions: exceptionTypeEnum | null;
  companyId?: string;
  communityId?: string;
}>;